perf: only calculate the distance between two particles once
Since distance is the same in both directions, the particles can be iterated over in pairs, and both their velocities can be computed at once. That way the expensive distance calculation only happens once per pair instead of twice when the particles are iterated over separately.

On my computer, if I disable the framerate limit this increases the fps from ~140 to ~220.